fix(io,linalg,ufunc,random): fix header validation, SVD convergence, reduction shapes, and RNG state width
Four correctness fixes improving NumPy parity:

1. fnp-io: Relax validate_required_header_keys to require at least the
   3 mandatory keys (descr, fortran_order, shape) instead of exactly 3,
   allowing valid .npy files with extra metadata keys like those
   produced by third-party tools. Update tests accordingly.

2. fnp-linalg: Add convergence tracking to svd_bidiag_full — if the
   maximum iteration budget is exhausted without convergence, return
   Err(LinAlgError::SvdNonConvergence) instead of silently returning
   potentially incorrect singular values.

3. fnp-ufunc: Remove the incorrect logic that forced shape [1] when
   reducing a 1D array along its only axis. NumPy produces a 0D scalar
   result, not a 1D array of length 1. Affects any/all/mean/sum/std
   and similar reduction operations. Also reject 0D arrays in
   where_nonzero with a ValueError matching NumPy's behavior.

4. fnp-random: Widen BitGeneratorState seed/counter fields from u64 to
   u128 to accommodate PCG64's full 128-bit state. Add bounded_u64
   zero-check to prevent division-by-zero. Add to_state_entries() for
   algorithm-specific state serialization. Relax schema validation to
   allow known algorithm-prefixed entries (mt19937_*, pcg64_*, etc.).

Also includes test_parity.py helper, conformance adjustments, and
UPGRADE_LOG / TODO documentation updates.

+++ crates/fnp-conformance/src/lib.rs
@@ -9607,7 +9607,7 @@ fn execute_rng_differential_operation(case: &RngDifferentialCase) -> Result<(), 
                 let _ = source.next_u64();
             }
             let (seed, counter) = source.state();
-            let mut restored = DeterministicRng::from_state(seed, counter);
+            let mut restored = DeterministicRng::from_state(u128::from(seed), u128::from(counter));
             for index in 0..replay_draws {
                 if source.next_u64() != restored.next_u64() {
                     return Err(RngSuiteError::new(
